use separate fragment shaders for each type
